feat: migrate marked text
