+++ src/themes/shared/slotRecipes/markedText.ts
@@ -54,4 +54,8 @@ export const markedTextRecipe = defineSlotRecipe({
       },
     },
   },
+
+  defaultVariants: {
+    variant: 'highlight',
+  },
 });
